update swap to handle new 'Swap success' status
